requirements.txt
Django==5.0.6
markdown2==2.4.13
requests==2.28.1
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
urllib3==1.26.10
